单测及部分修改

+++ springboot-test/springboot.service/src/main/java/com/daotong/springboot/service/domain/vo/LoadingVO.java
@@ -20,7 +20,7 @@ public class LoadingVO {
 
     private String loadingNo;
 
-    private Integer loadingId;
+    private Long loadingId;
 
     private Integer sendStationId;
